Move custom exception classes to a new `exceptions` package and update imports
