+++ src/main/java/sc/fiji/ome/zarr/pyramid/exceptions/NoMatchingResolutionException.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.ome.zarr.pyramid;
+package sc.fiji.ome.zarr.pyramid.exceptions;
 
 public class NoMatchingResolutionException extends RuntimeException
 {
