+++ src/main/java/sc/fiji/ome/zarr/pyramid/metadata/adapter/V04MetadataAdapter.java
@@ -35,7 +35,7 @@ import org.janelia.saalfeldlab.n5.universe.metadata.ome.ngff.v04.OmeNgffMetadata
 import org.janelia.saalfeldlab.n5.universe.metadata.ome.ngff.v04.OmeNgffMultiScaleMetadata;
 
 import net.imglib2.util.Cast;
-import sc.fiji.ome.zarr.pyramid.NotAMultiscaleImageException;
+import sc.fiji.ome.zarr.pyramid.exceptions.NotAMultiscaleImageException;
 import sc.fiji.ome.zarr.pyramid.metadata.Multiscale;
 
 class V04MetadataAdapter extends AbstractMetadataAdapter
