+++ src/test/java/sc/fiji/ome/zarr/pyramid/metadata/adapter/MetadataAdapterFactoryTest.java
@@ -50,7 +50,7 @@ import org.janelia.saalfeldlab.n5.universe.metadata.N5MetadataParser;
 import org.janelia.saalfeldlab.n5.universe.metadata.ome.ngff.v03.OmeNgffMetadata;
 import org.junit.jupiter.api.Test;
 
-import sc.fiji.ome.zarr.pyramid.NotAMultiscaleImageException;
+import sc.fiji.ome.zarr.pyramid.exceptions.NotAMultiscaleImageException;
 import sc.fiji.ome.zarr.pyramid.metadata.Multiscale;
 import sc.fiji.ome.zarr.util.ZarrTestUtils;
